Add server-side, async prop resolving lib
Created first draft of `ss-resolve` lib that accepts a
regular component and a promise based callback, and returns
a component that has markers recognizable on server that will
wait for them to resolve before sending response.

This allows container/route components to be written agnostically
of server-side requirements and instead can be passed into this
higher order function that converts it.