Add theme support via CSS variables
To use CSS variables with CSS modules, the path
of least resistence is to keep a theme CSS file
in the apps root and then @import it from styles
that need it.

Adjusted the webpack config to allow pathless @import
to pull from app's root.

Seeded theme file with color and media query variables.